README.md
The main reason I would use a template repository is so that all of the files I would typically use can be added to a template repository so that I don't have to add them each time you need to create a new repository. These files can be as simple as a customised .gitignore file, all the way up to a pre-defined template I use for an application.

Signed-off-by: Gara Takpara <davtakpara@hotmail.com>